refactor: file name change and some cleanup
re #1994
