+++ packages/taquito-core/src/taquito-core.ts
@@ -5,4 +5,4 @@
 
 export * from './error/errors';
 export * from './error/common';
-export * from './interface/status_code';
+export * from './interface/http_status_code';
